put SparseMatrix into namespace Protected

+++ trunk/linbox/linbox/algorithms/rational-solver.inl
@@ -1489,7 +1489,7 @@ namespace LinBox
 		FMatrix Ap(A, F);
 
 		// compute LQUP Factorization
-		Permutation<Field> P(A.coldim(),F),Q(A.rowdim(),F);
+		Permutation<Field> P((int)A.coldim(),F),Q((int)A.rowdim(),F);
 		FMatrix L(F, A.rowdim(), A.rowdim());
 		unsigned long rank;
 		Element_t det;
